[adc] Split files by platform (#7940)
